fix: fix question asking on suggested and ask ai button clicks
