+++ clients/search-component/src/TrieveModal/icons.tsx
@@ -6,7 +6,8 @@ export const LoadingIcon = (props: SVGAttributes<SVGElement>) => (
     width="120"
     height="30"
     viewBox="0 0 120 30"
-    {...props}>
+    {...props}
+  >
     <circle cx="15" cy="15" r="15" fill="currentColor">
       <animate
         attributeName="r"
@@ -16,7 +17,8 @@ export const LoadingIcon = (props: SVGAttributes<SVGElement>) => (
         from="15"
         repeatCount="indefinite"
         to="15"
-        values="15;9;15"></animate>
+        values="15;9;15"
+      ></animate>
       <animate
         attributeName="fill-opacity"
         begin="0s"
@@ -25,7 +27,8 @@ export const LoadingIcon = (props: SVGAttributes<SVGElement>) => (
         from="1"
         repeatCount="indefinite"
         to="1"
-        values="1;.5;1"></animate>
+        values="1;.5;1"
+      ></animate>
     </circle>
     <circle cx="60" cy="15" r="9" fillOpacity="0.3" fill="currentColor">
       <animate
@@ -36,7 +39,8 @@ export const LoadingIcon = (props: SVGAttributes<SVGElement>) => (
         from="9"
         repeatCount="indefinite"
         to="9"
-        values="9;15;9"></animate>
+        values="9;15;9"
+      ></animate>
       <animate
         attributeName="fill-opacity"
         begin="0s"
@@ -45,7 +49,8 @@ export const LoadingIcon = (props: SVGAttributes<SVGElement>) => (
         from="0.5"
         repeatCount="indefinite"
         to="0.5"
-        values=".5;1;.5"></animate>
+        values=".5;1;.5"
+      ></animate>
     </circle>
     <circle cx="105" cy="15" r="15" fill="currentColor">
       <animate
@@ -56,7 +61,8 @@ export const LoadingIcon = (props: SVGAttributes<SVGElement>) => (
         from="15"
         repeatCount="indefinite"
         to="15"
-        values="15;9;15"></animate>
+        values="15;9;15"
+      ></animate>
       <animate
         attributeName="fill-opacity"
         begin="0s"
@@ -65,7 +71,8 @@ export const LoadingIcon = (props: SVGAttributes<SVGElement>) => (
         from="1"
         repeatCount="indefinite"
         to="1"
-        values="1;.5;1"></animate>
+        values="1;.5;1"
+      ></animate>
     </circle>
   </svg>
 );
@@ -82,7 +89,8 @@ export const ArrowIcon = (props: SVGAttributes<SVGElement>) => {
       strokeWidth="2"
       strokeLinecap="round"
       strokeLinejoin="round"
-      {...props}>
+      {...props}
+    >
       <path stroke="none" d="M0 0h24v24H0z" fill="none" />
       <path d="M9 6l6 6l-6 6" />
     </svg>
@@ -101,7 +109,8 @@ export const AIIcon = (props: SVGAttributes<SVGElement>) => {
       strokeWidth="2"
       strokeLinecap="round"
       strokeLinejoin="round"
-      {...props}>
+      {...props}
+    >
       <path stroke="none" d="M0 0h24v24H0z" fill="none" />
       <path d="M16 18a2 2 0 0 1 2 2a2 2 0 0 1 2 -2a2 2 0 0 1 -2 -2a2 2 0 0 1 -2 2zm0 -12a2 2 0 0 1 2 2a2 2 0 0 1 2 -2a2 2 0 0 1 -2 -2a2 2 0 0 1 -2 2zm-7 12a6 6 0 0 1 6 -6a6 6 0 0 1 -6 -6a6 6 0 0 1 -6 6a6 6 0 0 1 6 6z" />
     </svg>
@@ -115,7 +124,8 @@ export const CheckIcon = (props: SVGAttributes<SVGElement>) => (
     height="24"
     viewBox="0 0 24 24"
     fill="currentColor"
-    {...props}>
+    {...props}
+  >
     <path stroke="none" d="M0 0h24v24H0z" fill="none" />
     <path d="M17 3.34a10 10 0 1 1 -14.995 8.984l-.005 -.324l.005 -.324a10 10 0 0 1 14.995 -8.336zm-1.293 5.953a1 1 0 0 0 -1.32 -.083l-.094 .083l-3.293 3.292l-1.293 -1.292l-.094 -.083a1 1 0 0 0 -1.403 1.403l.083 .094l2 2l.094 .083a1 1 0 0 0 1.226 0l.094 -.083l4 -4l.083 -.094a1 1 0 0 0 -.083 -1.32z" />
   </svg>
@@ -132,7 +142,8 @@ export const BackIcon = (props: SVGAttributes<SVGElement>) => (
     strokeWidth="2"
     strokeLinecap="round"
     strokeLinejoin="round"
-    {...props}>
+    {...props}
+  >
     <path stroke="none" d="M0 0h24v24H0z" fill="none" />
     <path d="M15 6l-6 6l6 6" />
   </svg>
@@ -145,10 +156,12 @@ export const LoadingAIIcon = (props: SVGAttributes<SVGElement>) => {
       height="23"
       viewBox="0 0 24 24"
       xmlns="http://www.w3.org/2000/svg"
-      {...props}>
+      {...props}
+    >
       <path
         d="M12,4a8,8,0,0,1,7.89,6.7A1.53,1.53,0,0,0,21.38,12h0a1.5,1.5,0,0,0,1.48-1.75,11,11,0,0,0-21.72,0A1.5,1.5,0,0,0,2.62,12h0a1.53,1.53,0,0,0,1.49-1.3A8,8,0,0,1,12,4Z"
-        fill="currentColor">
+        fill="currentColor"
+      >
         <animateTransform
           attributeName="transform"
           type="rotate"
@@ -167,7 +180,8 @@ export const EscKeyIcon = (props: SVGAttributes<SVGElement>) => (
       stroke="currentColor"
       strokeLinecap="round"
       strokeLinejoin="round"
-      strokeWidth="1.2">
+      strokeWidth="1.2"
+    >
       <path d="M13.6167 8.936c-.1065.3583-.6883.962-1.4875.962-.7993 0-1.653-.9165-1.653-2.1258v-.5678c0-1.2548.7896-2.1016 1.653-2.1016.8634 0 1.3601.4778 1.4875 1.0724M9 6c-.1352-.4735-.7506-.9219-1.46-.8972-.7092.0246-1.344.57-1.344 1.2166s.4198.8812 1.3445.9805C8.465 7.3992 8.968 7.9337 9 8.5c.032.5663-.454 1.398-1.4595 1.398C6.6593 9.898 6 9 5.963 8.4851m-1.4748.5368c-.2635.5941-.8099.876-1.5443.876s-1.7073-.6248-1.7073-2.204v-.4603c0-1.0416.721-2.131 1.7073-2.131.9864 0 1.6425 1.031 1.5443 2.2492h-2.956"></path>
     </g>
   </svg>
@@ -180,7 +194,8 @@ export const ArrowDownKey = (props: SVGAttributes<SVGElement>) => (
       stroke="currentColor"
       strokeLinecap="round"
       strokeLinejoin="round"
-      strokeWidth="1.2">
+      strokeWidth="1.2"
+    >
       <path d="M7.5 3.5v8M10.5 8.5l-3 3-3-3"></path>
     </g>
   </svg>
@@ -193,7 +208,8 @@ export const ArrowUpIcon = (props: SVGAttributes<SVGElement>) => (
       stroke="currentColor"
       strokeLinecap="round"
       strokeLinejoin="round"
-      strokeWidth="1.2">
+      strokeWidth="1.2"
+    >
       <path d="M7.5 11.5v-8M10.5 6.5l-3-3-3 3"></path>
     </g>
   </svg>
@@ -206,7 +222,8 @@ export const EnterKeyIcon = (props: SVGAttributes<SVGElement>) => (
       stroke="currentColor"
       strokeLinecap="round"
       strokeLinejoin="round"
-      strokeWidth="1.2">
+      strokeWidth="1.2"
+    >
       <path d="M12 3.53088v3c0 1-1 2-2 2H4M7 11.53088l-3-3 3-3"></path>
     </g>
   </svg>
@@ -223,7 +240,8 @@ export const ReloadIcon = (props: SVGAttributes<SVGElement>) => (
     strokeWidth="2"
     strokeLinecap="round"
     strokeLinejoin="round"
-    {...props}>
+    {...props}
+  >
     <path stroke="none" d="M0 0h24v24H0z" fill="none" />
     <path d="M19.933 13.041a8 8 0 1 1 -9.925 -8.788c3.899 -1 7.935 1.007 9.425 4.747" />
     <path d="M20 4v5h-5" />
@@ -233,12 +251,13 @@ export const ReloadIcon = (props: SVGAttributes<SVGElement>) => (
 export const UserIcon = (props: SVGAttributes<SVGElement>) => (
   <svg
     fill="currentColor"
-    stroke-width="0"
+    strokeWidth="0"
     xmlns="http://www.w3.org/2000/svg"
     viewBox="0 0 448 512"
     height="1em"
     width="1em"
-    {...props}>
+    {...props}
+  >
     <path d="M224 256a128 128 0 1 0 0-256 128 128 0 1 0 0 256zm-45.7 48C79.8 304 0 383.8 0 482.3 0 498.7 13.3 512 29.7 512h388.6c16.4 0 29.7-13.3 29.7-29.7 0-98.5-79.8-178.3-178.3-178.3h-91.4z"></path>
   </svg>
 );
